test: share the SDK integration scaffolding across the four suites
Each integration test carried its own copy of the mock connection, the
DDP frame helpers, and the collection/store builders. Move them into
app/lib/testUtils/sdkIntegration.ts and rename WireFrame to DdpMessage.
Jest mock registration stays per file and delegates to the shared
MockConnection through jest.requireActual so hoisting rules hold.
